Add the code for applying the patch to Virtuoso.
Also make sure not to try to import the "new.owl" file in the list of ontologies that have to be loaded by a vanilla instance.

We also changed the advancedIndexation parameter of the OntologyCreate endpoint from "false" to "true". That way, the ontologies will be imported in Virtuoso and Solr too.