Removed old Bash script related installation.
This branch is a development branch used to create the next generation of the installation script.

It will be using PHP as the command line scripting language, and it will have multiple other features. Amongst others, people will be able to only install and upgrade specific parts of the system. Everything will be modular.